feat: refactor SettingsClient component into smaller sections for email, password, and account deletion management; add themed components for consistent styling
